requirements.txt
Flask
pycryptodome
vaderSentiment
flask_pydantic
pydantic
